requirements.txt
pandas
PyPDF2
python-dotenv
openai
openpyxl
python-docx
colorama
tqdm